Refactored code into modules :package:
