+++ demo/ArcGisElcUI/index.html
@@ -22,9 +22,9 @@
                 async: true,
                 packages: [
                     {
-                        name: "elc",
+                        name: "RouteLocator",
                         location: root + "../../..",
-                        main: "elc"
+                        main: "RouteLocator"
                     },
                 ]
             };
